requirements.txt
docopt==0.6.1
PyYAML==3.10
requests==2.2.1
texttable==0.8.1
websocket-client==0.11.0
dockerpty==0.2.3
colander==1.0b1
